// front-end/ntution-front-end/src/app/models/institution.ts
export class institution {
    institutionName: string = "";
    streetAddr1: string = "";
    streetAddr2: string = "";
    country: string = "";
    state: string = "";
    city: string = "";
    postal: string = "";
    institutionType: string = "";
    privacyStatus: string = "";
    tuition_rates_instate: string = "";
    tuition_rates_outofstate: string = "";
    tuition_rates_online: string = "";
    tuition_rates_international: string = "";
}